Expose docker daemon with TLS
This commit adds infrastructure to support exposing the docker daemon
inside the ADB to the host.  This will allow tools on the host, such as
docker or eclipse, to access the docker daemon inside of the ADB.  This
is particularly useful on hosts which cannot directly support docker.

The commit consists of modifications to the startup of docker inside
of the ADB that generate TLS certificates, if they do not already exist.
The TLS certificates are copied into the vagrant users home directory
for potential later use by the vagrant-adbinfo plugin (see below).

The other component of this commit is that the docker daemon
configuration is modified to have the daemon listen on a TLS protected
TCP port and on the local unix socket.

The above changes are present in the kickstart file and as a separately
maintained script in build_tools/kickstarts/tls-certs-generation.sh

This change can standalone, but also works with the vagrant plugin
vagrant-adbinfo (https://github.com/bexelbie/vagrant-adbinfo).  This
plugin copies the TLS certificates from the ADB to the host and provides
the environment variables required to enable tools on the host to use
the ADB docker daemon.

Documentation updates are included, including a few minor typos.

This commit was authored by:
Navid Shaikh (nshaikh@redhat.com)
Brian Exelbierd (bex@pobox.com)